API audit (phase 7) + PlatformIO/Arduino/CubeIDE (phases 3-5)
- Start functions return uint8_t (1=accepted, 0=rejected); ds18b20_init -> uint8_t
- Add ds18b20_status_t enum, ds18b20_last_status(), ds18b20_set_callbacks(),
  ds18b20_deinit(); busy/complete dispatch via registered or weak callbacks
- ds18b20_busy(uint8_t action); onewire_search_stop()
- PlatformIO: library.json srcFilter excludes syscall.c, export.exclude
- Arduino: internal guards DS18B20_DRIVER_BUILD, syscall skips ARDUINO
- README/INTEGRATION signature updates; tests extended (harness_api)

+++ examples/app/app.c
@@ -330,7 +330,7 @@ void app_init(void) {
  *       Strong definition overrides the weak one in the DS18B20 driver.
  *       F1: LED on PC13 (active low). F0: LED on PA4 (active low assumed).
  */
-void ds18b20_busy(unsigned action) {
+void ds18b20_busy(uint8_t action) {
 #if defined(OW_PORT_FAMILY_F0) || defined(OW_PORT_FAMILY_G0)
     if (action) {
         // Turn LED on (PA4 low)
